STORE-2178 - Finished add tag to search result entry (selenium)
